+++ ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/tests/CommandService.scala
@@ -3,13 +3,10 @@
 
 package com.daml.ledger.api.testtool.tests
 
-import java.util.UUID
-
 import com.daml.ledger.api.testtool.infrastructure.Allocation._
 import com.daml.ledger.api.testtool.infrastructure.Assertions._
 import com.daml.ledger.api.testtool.infrastructure.TransactionHelpers._
 import com.daml.ledger.api.testtool.infrastructure.{LedgerSession, LedgerTestSuite}
-import com.digitalasset.ledger.api.v1.value.{Record, RecordField, Value}
 import com.digitalasset.ledger.client.binding.Primitive
 import com.digitalasset.ledger.client.binding.Value.encode
 import com.digitalasset.ledger.test_stable.Test.CallablePayout._
@@ -309,57 +306,6 @@ final class CommandService(session: LedgerSession) extends LedgerTestSuite(sessi
           failure,
           Status.Code.INVALID_ARGUMENT,
           "Command interpretation error in LF-DAMLe: Interpretation error: Error: User abort: Assertion failed. Details: Last location: [DA.Internal.Assert:20], partial transaction: root node"
-        )
-      }
-  }
-
-  test(
-    "CSExerciseByKey",
-    "Exercising by key should be possible only when the corresponding contract is available",
-    allocate(SingleParty)) {
-    case Participants(Participant(ledger, party)) =>
-      val keyString = UUID.randomUUID.toString
-      val expectedKey = Value(
-        Value.Sum.Record(
-          Record(
-            fields = Seq(
-              RecordField("_1", Some(Value(Value.Sum.Party(party.unwrap)))),
-              RecordField("_2", Some(Value(Value.Sum.Text(keyString))))
-            ))))
-      for {
-        failureBeforeCreation <- ledger
-          .exerciseByKey(
-            party,
-            TextKey.id,
-            expectedKey,
-            "TextKeyChoice",
-            Value(Value.Sum.Record(Record())))
-          .failed
-        _ <- ledger.create(party, TextKey(party, keyString, Primitive.List.empty))
-        _ <- ledger.exerciseByKey(
-          party,
-          TextKey.id,
-          expectedKey,
-          "TextKeyChoice",
-          Value(Value.Sum.Record(Record())))
-        failureAfterConsuming <- ledger
-          .exerciseByKey(
-            party,
-            TextKey.id,
-            expectedKey,
-            "TextKeyChoice",
-            Value(Value.Sum.Record(Record())))
-          .failed
-      } yield {
-        assertGrpcError(
-          failureBeforeCreation,
-          Status.Code.INVALID_ARGUMENT,
-          "dependency error: couldn't find key"
-        )
-        assertGrpcError(
-          failureAfterConsuming,
-          Status.Code.INVALID_ARGUMENT,
-          "dependency error: couldn't find key"
         )
       }
   }
